+++ js/models/message_handler.js
@@ -132,6 +132,13 @@ export default class MessageHandler {
     }
   }
 
+  handleRemoveOverlay = (msg) => {
+    const overlay_names = msg["overlay_names"];
+    for (const overlay_name of overlay_names) {
+      this.aladin.removeOverlay(overlay_name);
+    }
+  };
+
   handleChangeColormap(msg) {
     this.aladin.getBaseImageLayer().setColormap(msg["colormap"]);
   }