+++ js/aladin_lite.js
@@ -1,3 +1,3 @@
-import A from "https://esm.sh/aladin-lite@3.7.3-beta";
+import A from "https://esm.sh/aladin-lite@3.9.0-beta";
 
 export default A;