+++ src/ipyaladin/__about__.py
@@ -1,2 +1,2 @@
-__version__ = "0.7.0"
-__aladin_lite_version__ = "3.7.3-beta"
+__version__ = "0.8.0"
+__aladin_lite_version__ = "3.9.0-beta"