+++ CHANGELOG.md
@@ -16,6 +16,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- `projection` parameter can now be called to change the projection of the
+  ipyaladin viewer [#172]
+
 ## [0.7.0]
 
 ## Added